Don't use inline assets for built-in kinos
